Fix layout on forms

+++ public/js/expense.js
@@ -17,6 +17,16 @@ $(`#category [value=${categoryId}]`).attr("selected", "selected");
 const expenseId = $("#iddata").val();
 console.log(expenseId);
 
+// Convert amount to fixed 2 decimals when focusedout.
+$("#amount").on("focusout", function(event) {
+  event.preventDefault();
+
+  let inputVal = parseFloat($(this).val());
+  if (inputVal) {
+      $(this).val(inputVal.toFixed(2));
+  }
+});
+
 
 // When the edit form is submitted
 $("#edit-expense").on("submit", function(event) {
@@ -40,7 +50,16 @@ $("#edit-expense").on("submit", function(event) {
     type: "PUT",
     data: updatedExpense
   }).then(function(dbExpense) {
-    location.reload();
+    // location.reload();
+    $("#message-section").html(`
+      <div class="alert alert-primary alert-dismissible" role="alert">
+        ${updatedExpense.description} was updated!
+
+        <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+    `)
   });
 });
 
@@ -49,10 +68,23 @@ $("#edit-expense").on("submit", function(event) {
 $("#delete").on("click", function(event) {
   event.preventDefault();
 
-  // Do the ajax PUT request
+  if (!confirm("Are you sure you want to delete this?")) {
+    return;
+  }
+
+  // Do the ajax DELETE request
   $.ajax("/api/expenses/" + expenseId, {
     type: "DELETE"
   }).then(function(dbExpense) {
-    location.reload();
+    // location.reload();
+    $("#message-section").html(`
+      <div class="alert alert-danger alert-dismissible" role="alert">
+        ${$("#description").val().trim()} was deleted!
+
+        <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+          <span aria-hidden="true">&times;</span>
+        </button>
+      </div>
+    `)
   });
 })
